Fixed submission save, edit cancel
Related Work Items: #4908

+++ ModernSlavery.BusinessDomain.Submission/StatementBusinessLogic.cs
@@ -535,8 +535,6 @@ namespace ModernSlavery.BusinessDomain.Submission
             //Get the statement to map from
             var submittedStatement = await FindSubmittedStatementAsync(organisation.OrganisationId, reportingDeadline);
 
-            var createBackup = false;
-
             if (draftStatement != null)
             {
                 //Check if the existing draft lock has expired
@@ -545,8 +543,6 @@ namespace ModernSlavery.BusinessDomain.Submission
                 //Check the existing draft is not still locked by another user
                 if (draftStatement.EditorUserId > 0 && draftStatement.EditorUserId != userId && !draftExpired)
                     return new Outcome<StatementErrors, StatementModel>(StatementErrors.Locked);
-
-                createBackup = submittedStatement==null;
             }
             else
             {
@@ -558,7 +554,6 @@ namespace ModernSlavery.BusinessDomain.Submission
                     //Create an empty draft statement
                     draftStatement = _mapper.Map<StatementModel>(GetEmptyStatement(organisation, reportingDeadline));
                     //Create a backup for non-submitted drafts
-                    createBackup = true;
                 }
             }
 
@@ -572,7 +567,7 @@ namespace ModernSlavery.BusinessDomain.Submission
             draftStatement.Submitted = organisation.Statements.Any(s => s.Status == StatementStatuses.Submitted && s.SubmissionDeadline == reportingDeadline);
 
             //Save new statement model with timestamp to lock to new user
-            await SaveDraftStatementModelAsync(draftStatement, createBackup);
+            await SaveDraftStatementModelAsync(draftStatement, true);
             draftStatement.DraftBackupDate = null;
 
             //Get the draft backup date
@@ -607,24 +602,42 @@ namespace ModernSlavery.BusinessDomain.Submission
             //Get the backup draft filepath
             var draftBackupFilePath = GetDraftBackupFilepath(organisation.OrganisationId, reportingDeadline.Year);
 
-            var backupStatementModel = await LoadStatementModelFromFile(draftBackupFilePath);
-            if (backupStatementModel != null)
-            {
-                //Restore the original draft from the backup unlkess its empty
-                if (!backupStatementModel.IsEmpty())
-                    await _sharedBusinessLogic.FileRepository.CopyFileAsync(draftBackupFilePath, draftFilePath, true);
+            //Remove the backup if its empty or submitted
+            var backupExists=await DeleteIfEmptyOrSubmittedAsync(draftBackupFilePath);
 
-                //Delete the backup draft
+            //Restore draft from backup
+            if (backupExists)
+            { 
+                await _sharedBusinessLogic.FileRepository.CopyFileAsync(draftBackupFilePath, draftFilePath, true);
                 await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftBackupFilePath);
             }
 
-            else if (await _sharedBusinessLogic.FileRepository.GetFileExistsAsync(draftFilePath))
-            {
-                //Delete the draft
-                await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftFilePath);
-            }
+            //Remove the draft if its empty or submitted
+            await DeleteIfEmptyOrSubmittedAsync(draftFilePath);
 
             return new Outcome<StatementErrors>();
+        }
+
+        /// <summary>
+        /// Deletes a draft file if it is empty or same as last submitted
+        /// </summary>
+        /// <param name="draftFilePath"></param>
+        /// <returns>True if the file remains</returns>
+        private async Task<bool> DeleteIfEmptyOrSubmittedAsync(string draftFilePath, StatementModel statementModel=null)
+        {
+            if (statementModel==null)statementModel = await LoadStatementModelFromFile(draftFilePath);
+            if (statementModel == null) return false;
+            
+            var delete = statementModel.IsEmpty();
+            if (!delete && statementModel.Submitted)
+            {
+                statementModel.Modifications = await CompareToSubmittedStatement(statementModel);
+                delete = !statementModel.Modifications.Any();
+            }
+            //Delete the draft
+            if (delete) await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftFilePath);
+
+            return !delete;
         }
 
         public async Task<Outcome<StatementErrors>> CloseDraftStatementModelAsync(long organisationId, int reportingDeadlineYear, long userId)
@@ -641,30 +654,26 @@ namespace ModernSlavery.BusinessDomain.Submission
             //Validate the parameters
             if (userId <= 0) throw new ArgumentOutOfRangeException(nameof(userId));
 
+            //Get the draft
             var openOutcome = await OpenDraftStatementModelAsync(organisation, reportingDeadline, userId);
             if (openOutcome.Fail) return new Outcome<StatementErrors>(openOutcome.Errors);
-
             var statementModel = openOutcome.Result;
-            statementModel.EditorUserId = 0;
 
-            //Get the current draft filepath
+            //Remove the draft if its empty or submitted
             var draftFilePath = GetDraftFilepath(statementModel.OrganisationId, statementModel.SubmissionDeadline.Year);
+            var draftExists = await DeleteIfEmptyOrSubmittedAsync(draftFilePath, statementModel);
 
-            if (statementModel.IsEmpty())
-                //Delete the empty draft
-                await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftFilePath);
-            else
-                //Save the draft with no userId
-                await SaveStatementModelToFileAsync(statementModel, draftFilePath);
-
-            //Get the backup draft filepath
-            var draftBackupFilePath = GetDraftBackupFilepath(statementModel.OrganisationId, reportingDeadline.Year);
-
-            if (await _sharedBusinessLogic.FileRepository.GetFileExistsAsync(draftBackupFilePath))
+            //Save the draft with no userId
+            if (draftExists)
             {
-                //Delete the backup draft
-                await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftBackupFilePath);
+                statementModel.EditorUserId = 0;
+                await SaveStatementModelToFileAsync(statementModel, draftFilePath);
             }
+
+            //Delete the backup draft
+            var draftBackupFilePath = GetDraftBackupFilepath(statementModel.OrganisationId, reportingDeadline.Year);
+            if (await _sharedBusinessLogic.FileRepository.GetFileExistsAsync(draftBackupFilePath))
+                await _sharedBusinessLogic.FileRepository.DeleteFileAsync(draftBackupFilePath);
 
             return new Outcome<StatementErrors>();
         }
